feat(repeater): address pr comments
closes #196
