+++ packages/repeater/src/api/index.ts
@@ -5,6 +5,4 @@ export {
   RegisterRepeaterCommand
 } from './commands';
 export { RepeaterStatusEvent } from './events';
-export * from './RepeatersManager';
-export * from './DefaultRepeatersManager';
 export * from './ExecuteRequestEventHandler';
