+++ packages/repeater/src/bus/RepeaterBus.ts
@@ -1,4 +1,5 @@
 export interface RepeaterBus {
+  readonly repeaterId?: string;
   connect(): Promise<void>;
   close(): Promise<void>;
 }
