+++ packages/repeater/src/bus/RepeaterBusFactory.ts
@@ -1,7 +1,7 @@
 import { RepeaterBus } from './RepeaterBus';
 
 export interface RepeaterBusFactory {
-  create(repeaterId: string): RepeaterBus;
+  create(): RepeaterBus;
 }
 
 export const RepeaterBusFactory: unique symbol = Symbol('RepeaterBusFactory');
