--- packages/repeater/src/bus/RepeaterCommandHub.ts
@@ -1,7 +1,0 @@
-import { Request, Response } from '../request-runner';
-
-export interface RepeaterCommandHub {
-  sendRequest(request: Request): Promise<Response>;
-}
-
-export const RepeaterCommandHub: unique symbol = Symbol('RepeaterCommandHub');
