+++ packages/repeater/src/bus/RepeaterCommands.ts
@@ -1,0 +1,7 @@
+import { Request, Response } from '../request-runner';
+
+export interface RepeaterCommands {
+  sendRequest(request: Request): Promise<Response>;
+}
+
+export const RepeaterCommands: unique symbol = Symbol('RepeaterCommands');
