+++ packages/repeater/src/bus/index.ts
@@ -1,9 +1,7 @@
 export * from './DefaultRepeaterBusFactory';
-export * from './DefaultRepeaterCommandHub';
-export * from './DefaultRepeaterEventHub';
+export * from './DefaultRepeaterCommands';
 export * from './DefaultRepeaterServer';
 export * from './RepeaterBus';
 export * from './RepeaterBusFactory';
-export * from './RepeaterCommandHub';
-export * from './RepeaterEventHub';
+export * from './RepeaterCommands';
 export * from './RepeaterServer';
