+++ packages/repeater/src/lib/Repeater.spec.ts
@@ -8,13 +8,10 @@ describe('Repeater', () => {
   let repeater!: Repeater;
   const mockedRepeaterBus = mock<RepeaterBus>();
 
-  const createRepeater = () =>
-    new Repeater({
-      repeaterId,
-      bus: instance(mockedRepeaterBus)
-    });
+  const createRepeater = () => new Repeater(instance(mockedRepeaterBus));
 
   beforeEach(() => {
+    when(mockedRepeaterBus.repeaterId).thenReturn(repeaterId);
     repeater = createRepeater();
   });
