+++ packages/repeater/src/lib/RepeaterFactory.spec.ts
@@ -7,11 +7,9 @@ import {
   RequestRunnerOptions
 } from '../request-runner';
 import { Repeater } from './Repeater';
-import { RepeatersManager } from '../api';
 import { Configuration } from '@sectester/core';
 import {
   anything,
-  capture,
   deepEqual,
   instance,
   mock,
@@ -23,7 +21,6 @@ import {
 import { DependencyContainer, Lifecycle } from 'tsyringe';
 
 describe('RepeaterFactory', () => {
-  const repeaterId = 'fooId';
   const defaultOptions = {
     timeout: 30000,
     maxContentLength: 100,
@@ -50,7 +47,6 @@ describe('RepeaterFactory', () => {
   const mockedChildContainer = mock<DependencyContainer>();
   const mockedConfiguration = mock<Configuration>();
   const mockedRepeaterBus = mock<RepeaterBus>();
-  const mockedRepeaterManager = mock<RepeatersManager>();
   const mockedRepeaterBusFactory = mock<RepeaterBusFactory>();
 
   const configuration = instance(mockedConfiguration);
@@ -65,36 +61,23 @@ describe('RepeaterFactory', () => {
     );
 
     when(
-      mockedContainer.resolve<RepeatersManager>(RepeatersManager)
-    ).thenReturn(instance(mockedRepeaterManager));
-
-    when(mockedRepeaterManager.createRepeater(anything())).thenResolve({
-      repeaterId
-    });
-
-    when(
       mockedChildContainer.resolve<RepeaterBusFactory>(RepeaterBusFactory)
     ).thenReturn(instance(mockedRepeaterBusFactory));
 
-    when(mockedRepeaterBusFactory.create(repeaterId)).thenReturn(
+    when(mockedRepeaterBusFactory.create()).thenReturn(
       instance(mockedRepeaterBus)
     );
   });
 
   afterEach(() => {
     reset<
-      | DependencyContainer
-      | Configuration
-      | RepeaterBus
-      | RepeatersManager
-      | RepeaterBusFactory
+      DependencyContainer | Configuration | RepeaterBus | RepeaterBusFactory
     >(
       mockedContainer,
       mockedChildContainer,
       mockedConfiguration,
       mockedRepeaterBus,
-      mockedRepeaterBusFactory,
-      mockedRepeaterManager
+      mockedRepeaterBusFactory
     );
   });
 
@@ -107,66 +90,6 @@ describe('RepeaterFactory', () => {
       const res = await factory.createRepeater();
 
       // assert
-      expect(res).toBeInstanceOf(Repeater);
-      expect(res).toMatchObject({
-        repeaterId
-      });
-    });
-
-    it('should create repeater with given name prefix and description', async () => {
-      // arrange
-      const factory = new RepeaterFactory(configuration);
-
-      // act
-      const res = await factory.createRepeater({
-        namePrefix: 'foo',
-        description: 'description'
-      });
-
-      const [arg]: [
-        {
-          name: string;
-          description?: string;
-        }
-      ] = capture<{
-        name: string;
-        description?: string;
-      }>(mockedRepeaterManager.createRepeater).first();
-
-      // assert
-      expect(arg?.name).toMatch(/^foo/);
-      expect(arg?.description).toBe('description');
-      expect(res).toBeInstanceOf(Repeater);
-    });
-
-    it('should create repeater with given name without the random postfix', async () => {
-      // arrange
-      const factory = new RepeaterFactory(configuration);
-
-      // act
-      const res = await factory.createRepeater({
-        namePrefix: 'foo',
-        disableRandomNameGeneration: true
-      });
-
-      // assert
-      verify(
-        mockedRepeaterManager.createRepeater(objectContaining({ name: 'foo' }))
-      );
-      expect(res).toBeInstanceOf(Repeater);
-    });
-
-    it('should create repeater with given project', async () => {
-      // arrange
-      const factory = new RepeaterFactory(configuration);
-      const projectId = '321';
-      const res = await factory.createRepeater({
-        projectId
-      });
-
-      verify(
-        mockedRepeaterManager.createRepeater(objectContaining({ projectId }))
-      );
       expect(res).toBeInstanceOf(Repeater);
     });
 
@@ -243,37 +166,6 @@ describe('RepeaterFactory', () => {
           })
         )
       ).once();
-    });
-
-    it('should throw an error if name prefix is too long', async () => {
-      // arrange
-      const factory = new RepeaterFactory(configuration);
-
-      // act
-      const res = factory.createRepeater({
-        namePrefix: 'foo'.repeat(50)
-      });
-
-      // assert
-      await expect(res).rejects.toThrow(
-        'Name prefix must be less than or equal to 43 characters.'
-      );
-    });
-
-    it('should throw an error when name prefix is too long and random postfix is disabled', async () => {
-      // arrange
-      const factory = new RepeaterFactory(configuration);
-
-      // act
-      const res = factory.createRepeater({
-        namePrefix: 'foo'.repeat(80),
-        disableRandomNameGeneration: true
-      });
-
-      // assert
-      await expect(res).rejects.toThrow(
-        'Name prefix must be less than or equal to 80 characters.'
-      );
     });
 
     it('should throw an error when credentials was not provided', async () => {
