+++ packages/repeater/src/lib/RepeaterFactory.ts
@@ -1,12 +1,9 @@
-import { Repeater, RepeaterId } from './Repeater';
+import { Repeater } from './Repeater';
 import { RequestRunner, RequestRunnerOptions } from '../request-runner';
 import { RepeaterOptions } from './RepeaterOptions';
-import { RepeatersManager } from '../api';
 import { RepeaterBusFactory } from '../bus/RepeaterBusFactory';
 import { DefaultRepeaterServerOptions } from '../bus/DefaultRepeaterServer';
-import { RepeaterRequestRunnerOptions } from './RepeaterRequestRunnerOptions';
 import { Configuration } from '@sectester/core';
-import { v4 as uuidv4 } from 'uuid';
 import { DependencyContainer, injectable, Lifecycle } from 'tsyringe';
 
 /**
@@ -14,47 +11,18 @@ import { DependencyContainer, injectable, Lifecycle } from 'tsyringe';
  */
 @injectable()
 export class RepeaterFactory {
-  private readonly MAX_NAME_LENGTH = 80;
-  private readonly repeatersManager: RepeatersManager;
   private readonly runnerOptions: Readonly<RequestRunnerOptions>;
 
   constructor(private readonly configuration: Configuration) {
-    this.repeatersManager =
-      this.configuration.container.resolve(RepeatersManager);
     this.runnerOptions =
       this.configuration.container.resolve(RequestRunnerOptions);
   }
 
   public async createRepeater({
-    projectId,
-    description,
-    disableRandomNameGeneration,
-    namePrefix = 'sectester',
-    ...requestRunnerOptions
+    requestRunnerOptions,
+    requestRunners = []
   }: RepeaterOptions = {}): Promise<Repeater> {
-    const name = this.generateName(namePrefix, disableRandomNameGeneration);
-
-    const { repeaterId } = await this.repeatersManager.createRepeater({
-      description,
-      projectId,
-      name
-    });
-
-    return this.createRepeaterInstance(repeaterId, requestRunnerOptions);
-  }
-
-  private async createRepeaterInstance(
-    repeaterId: string,
-    {
-      requestRunnerOptions,
-      requestRunners = []
-    }: RepeaterRequestRunnerOptions = {}
-  ): Promise<Repeater> {
     const container = this.configuration.container.createChildContainer();
-
-    container.register(RepeaterId, {
-      useValue: repeaterId
-    });
 
     await this.registerRepeaterServerOptions(container);
     this.registerRequestRunnerOptions(container, requestRunnerOptions);
@@ -63,10 +31,7 @@ export class RepeaterFactory {
     const busFactory =
       container.resolve<RepeaterBusFactory>(RepeaterBusFactory);
 
-    return new Repeater({
-      repeaterId,
-      bus: busFactory.create(repeaterId)
-    });
+    return new Repeater(busFactory.create());
   }
 
   private async registerRepeaterServerOptions(
@@ -90,25 +55,6 @@ export class RepeaterFactory {
         }
       }
     );
-  }
-
-  private generateName(
-    namePrefix: string,
-    disableRandomNameGeneration: boolean = false
-  ) {
-    const normalizedPrefix = namePrefix?.trim();
-    const randomPostfix = disableRandomNameGeneration ? '' : `-${uuidv4()}`;
-    const name = `${normalizedPrefix}${randomPostfix}`;
-
-    if (name.length > this.MAX_NAME_LENGTH) {
-      const maxPrefixLength = this.MAX_NAME_LENGTH - randomPostfix.length;
-
-      throw new Error(
-        `Name prefix must be less than or equal to ${maxPrefixLength} characters.`
-      );
-    }
-
-    return name;
   }
 
   private registerRequestRunners(
