+++ packages/repeater/src/lib/index.ts
@@ -1,4 +1,4 @@
-export { Repeater, RunningStatus, RepeaterId } from './Repeater';
+export { Repeater, RunningStatus } from './Repeater';
 export { RepeaterFactory } from './RepeaterFactory';
 export { RepeaterOptions } from './RepeaterOptions';
 export { RepeaterRequestRunnerOptions } from './RepeaterRequestRunnerOptions';
