+++ packages/runner/src/lib/SecRunner.ts
@@ -1,11 +1,7 @@
 import { SecScanOptions } from './SecScanOptions';
 import { SecScan } from './SecScan';
 import { Configuration, ConfigurationOptions, Logger } from '@sectester/core';
-import {
-  Repeater,
-  RepeaterFactory,
-  RepeatersManager
-} from '@sectester/repeater';
+import { Repeater, RepeaterFactory } from '@sectester/repeater';
 import { ScanFactory } from '@sectester/scan';
 import { Formatter, PlainTextFormatter } from '@sectester/reporter';
 
@@ -19,7 +15,6 @@ export class SecRunner {
   private readonly logger: Logger;
   private repeater: Repeater | undefined;
   private repeaterFactory: RepeaterFactory | undefined;
-  private repeatersManager: RepeatersManager | undefined;
 
   get repeaterId(): string | undefined {
     return this.repeater?.repeaterId;
@@ -32,14 +27,12 @@ export class SecRunner {
   }
 
   public async init(): Promise<void> {
-    if (this.repeatersManager && this.repeaterFactory) {
+    if (this.repeaterFactory) {
       throw new Error('Already initialized.');
     }
 
     await this.initConfiguration(this.configuration);
 
-    this.repeatersManager =
-      this.configuration.container.resolve(RepeatersManager);
     this.repeaterFactory =
       this.configuration.container.resolve(RepeaterFactory);
 
@@ -52,14 +45,12 @@ export class SecRunner {
 
   public async clear(): Promise<void> {
     try {
-      if (this.repeater && this.repeatersManager) {
+      if (this.repeater) {
         await this.repeater.stop();
-        await this.repeatersManager.deleteRepeater(this.repeater.repeaterId);
       }
     } finally {
       this.removeShutdownHandler();
       delete this.repeater;
-      delete this.repeatersManager;
       delete this.repeaterFactory;
     }
   }
